@@ -1,53 +1,56 @@
 {
   "name": "turn-based-accounting",
-  "version": "2.0.4",
+  "version": "2.0.5",
   "main": "index.js",
   "license": "MIT",
   "sideEffects": false,
+  "type": "module",
   "scripts": {
     "dev": "cross-env NODE_ENV=development remix dev",
     "build": "remix build",
-    "start": "remix-serve build",
+    "start": "remix-serve ./build/index.js",
     "typecheck": "tsc",
     "test": "vitest"
   },
   "prisma": {
-    "seed": "node --require esbuild-register prisma/seed.ts"
+    "seed": "node --loader ts-node/esm prisma/seed.ts"
   },
   "dependencies": {
-    "@headlessui/react": "1.7.14",
-    "@heroicons/react": "2.0.18",
-    "@prisma/client": "4.14.1",
-    "@remix-run/node": "1.16.1",
-    "@remix-run/react": "1.16.1",
-    "@remix-run/serve": "1.16.1",
-    "clsx": "1.2.1",
-    "date-fns": "2.30.0",
-    "isbot": "3.6.10",
-    "jsonwebtoken": "9.0.0",
+    "@headlessui/react": "1.7.18",
+    "@heroicons/react": "2.1.1",
+    "@prisma/client": "5.8.1",
+    "@remix-run/node": "2.5.1",
+    "@remix-run/react": "2.5.1",
+    "@remix-run/serve": "2.5.1",
+    "clsx": "2.1.0",
+    "date-fns": "3.3.1",
+    "isbot": "^4",
+    "jsonwebtoken": "9.0.2",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "remix-typedjson": "0.1.7",
+    "remix-typedjson": "0.4.1",
     "tiny-invariant": "1.3.1"
   },
   "devDependencies": {
-    "@faker-js/faker": "8.0.1",
-    "@remix-run/dev": "1.16.1",
-    "@remix-run/eslint-config": "1.16.1",
-    "@tailwindcss/forms": "0.5.3",
-    "@types/jsonwebtoken": "9.0.2",
-    "@types/react": "18.2.7",
-    "@types/react-dom": "18.2.4",
-    "@vitejs/plugin-react": "4.0.0",
-    "babel-loader": "9.1.2",
+    "@faker-js/faker": "8.4.0",
+    "@remix-run/dev": "2.5.1",
+    "@remix-run/eslint-config": "2.5.1",
+    "@tailwindcss/forms": "0.5.7",
+    "@types/jsonwebtoken": "9.0.5",
+    "@types/node": "^20.11.10",
+    "@types/react": "18.2.48",
+    "@types/react-dom": "18.2.18",
+    "@vitejs/plugin-react": "4.2.1",
+    "babel-loader": "9.1.3",
     "cross-env": "7.0.3",
-    "esbuild-register": "3.4.2",
-    "eslint": "8.41.0",
-    "jsdom": "22.0.0",
-    "prisma": "4.14.1",
-    "tailwindcss": "3.3.2",
-    "typescript": "5.0.4",
-    "vitest": "0.31.1"
+    "esbuild-register": "3.5.0",
+    "eslint": "8.56.0",
+    "jsdom": "24.0.0",
+    "prisma": "5.8.1",
+    "tailwindcss": "3.4.1",
+    "ts-node": "^10.9.2",
+    "typescript": "5.3.3",
+    "vitest": "1.2.2"
   },
   "engines": {
     "node": ">=18"
