@@ -1,15 +1,9 @@
 /**
  * @type {import('@remix-run/dev').AppConfig}
  */
-module.exports = {
+export default {
   ignoredRouteFiles: ["**/.*"],
   tailwind: true,
-  // serverModuleFormat: "esm",
   future: {
-    unstable_dev: false,
-    v2_routeConvention: true,
-    v2_meta: true,
-    v2_errorBoundary: true,
-    v2_normalizeFormMethod: true,
   },
 };
